sqs amazon/requirements.txt
python-dotenv
psycopg2
sqlalchemy
boto3
sqlalchemy_utils